pcb_pipeline/01_load_and_inspect.py
import os
import glob
import cv2
import matplotlib.pyplot as plt
from utils import DATA_DIR, OUTPUT_DIR, ensure_dirs, save_json, parse_label_from_filename, get_binary_label

def main():
    ensure_dirs()
    
    image_files = glob.glob(os.path.join(DATA_DIR, "*.png"))
    if not image_files:
        print(f"Error: No PNG images found in {DATA_DIR}")
        return
    
    print(f"Found {len(image_files)} image files. Inspecting...")
    
    metadata = {}
    ok_count = 0
    ng_count = 0
    defect_types = set()
    
    for filepath in image_files:
        filename = os.path.basename(filepath)
        file_size = os.path.getsize(filepath)
        
        # We don't read full images just for size to speed things up, 
        # but the requirements asked to inspect them. 
        # Using PIL just for dimension extraction is much faster than cv2.imread for 4k images.
        from PIL import Image
        with Image.open(filepath) as img:
            width, height = img.size
            mode = img.mode
            
        label = parse_label_from_filename(filename)
        binary_label = get_binary_label(label)
        
        if binary_label == "clean":
            ok_count += 1
        else:
            ng_count += 1
            defect_types.add(label)
            
        metadata[filename] = {
            "filename": filename,
            "path": filepath,
            "width": width,
            "height": height,
            "mode": mode,
            "size_bytes": file_size,
            "label": label,
            "binary_label": binary_label
        }
        
    metadata_path = os.path.join(OUTPUT_DIR, "metadata.json")
    save_json(metadata_path, metadata)
    
    print("\n--- Summary ---")
    print(f"Total Images: {len(image_files)}")
    print(f"Clean (OK) Images: {ok_count}")
    print(f"Faulty (NG) Images: {ng_count}")
    print(f"Defect Types: {sorted(list(defect_types))}")
    print(f"Metadata saved to: {metadata_path}")
    print("----------------")
    
    # Optional grid plotting removed for headless compatibility during test runs,
    # but can be added back if needed by user. Since step 2 requires interaction anyway,
    # we'll provide a plot option.
    
    print("Done inspecting data.")

if __name__ == "__main__":
    main()
