pcb_pipeline/requirements.txt
numpy
opencv-python-headless
matplotlib
PyQt5
scikit-learn
joblib
Pillow
scikit-image
